reorganize DI file structure
