+++ lib/backend/interfaces/ICharacterService.ts
@@ -1,9 +1,7 @@
-import { Character, PrismaClient } from ".prisma/client";
+import { Character } from ".prisma/client";
 
 export interface ICharacterService {
     createCharacter(character: Character): Promise<Character>;
     getAllCharacters(): Promise<Character[]>;
     getCharacterById(id: number): Promise<Character>;
 }
-
-export type Prisma = PrismaClient;
