+++ lib/backend/interfaces/Prisma.ts
@@ -1,0 +1,3 @@
+import { PrismaClient } from ".prisma/client";
+
+export type Prisma = PrismaClient;
