+++ lib/backend/interfaces/index.ts
@@ -1,0 +1,4 @@
+import { ICharacterService } from "./ICharacterService";
+import { Prisma } from "./Prisma";
+
+export type { ICharacterService, Prisma };
