+++ lib/backend/services/index.ts
@@ -1,0 +1,5 @@
+import CharacterService from "./CharacterService";
+
+export {
+    CharacterService
+};
